moved tree into the analysis module, started StateNetwork
Updated functions in WepyHDF5 to work with network stuff. Mainly this
is having traces which also have the run idx in them.

Moved teh tree.ancestors to the new submodule analysis.tree

started a new analysis.network submodule with the class StateNetwork
which is going to be used for conformation state networks.

Right now you can get fields from it using the state labels.

I want to add edge support later, but this might require integrating
with networkx?