mapper, worker, and GPU worker modules; examples of mappers
I made three examples showing the different types of work mappers that
I have been working on in the Lennard Jones system:

- gpu_run.py :: showing how to run on the GPU for OpenMM
- serial_run.py :: a simple builtin map mapper that runs in serial
- workmapper_run.py :: uses Processes and the reference platform

Moved the old GPU mapper to the prototypes.

saved the prototype for the new work mapper design I have been working
on, called example4.py

Added the OpenMMGPUWorker class to runners.openmm.py which is all that
is needed given the framework in the new mapper and worker modules.

Made some updates and changes to the behavior of sim_manager.
1. instead of taking a work_mapper object it takes the class and the
worker type the work_mapper is to be initialized with. This is because
that multiprocessing applications required that the function being
mapped be defined above the __main__ block thus it was necessary to
pass it in as an argument to the constructor for the work mapper
itself, instead of giving it to every invocation of map, which would
occur in the __main__ block.

This makes the invocation to the sim_manager more verbose
unfortunately, although this could be ameliorated with a bundled
OpenMM sim manager at the application level.

2. Allowed None values for many of the inputs that don't need to be
there, like reporters and the worker_type, which has a default in the
work mapper.

3. The function signature of the work_mapper.map has thus been changed
so that it should no longer accept a function. Just arguments for
computing the function; and kwargs that are intended (as of now and
this may change) to be for use in signalling from the sim_manager to
the mapper.

Thus this really only effects the API for the work mapper and not the
sim_manager.

Added the work_mapper.mapper module:
- Mapper :: basic serial mapper that uses the python builtin map
- WorkerMapper :: uses multiprocessing to make new processes for the
workers and allows mapping of the workers to Tasks. Has methods that
can be inherited and overriden:
  - make_task :: which can be customized to pass in Mapper specific
  data
  - map :: according to the original interface API for wepy sim_manager

Complementary to this module is the Worker module:
- Worker :: subclass of Process that while running takes things off of
the task Queue and runs them. Methods:
  - run_task :: which can be overriden to pass in Worker specific data
  such as the worker idx etc. as is done in the OpenMMGPUWorker
  - run :: necessary for the multiprocessing API for a process.
- Task :: simple class that stores a function and the inputs to be
run, which can be scheduled or consumed by a worker and then run.

Next steps will be to alter the way processes are created in the
WorkerMapper so that they are only created once whent the class is
initialized.

This should also take place not in the sim_manager constructor as it
is now, but in the run_simulation.

THis architecture would allow for tracking of specific pieces of
hardware across simulations in case errors arise or to allow us to put
timeouts on devices in case they stall etc.

More data on the identity of the GPUs could be gotten from the nvidia
python package which have used before.