vlen aux_data, unbinding min_distance dtype
WARNING this commit may break some other implementations of boundary
conditions!!

Please read to understand why.

I wanted to be able to have the min_distance specified so the HDF5
could accomodate for it properly.

THis involved shaving a yak.

So now you can specify the shape for an aux data type as Ellipsis
which will make it a variable length dtype using the special dtype in
h5py.

THis required special handling because these are different shapes than
the rest of the data that is given to aux_data.

It looks the same coming out, almost (so it will require some checking
but hopefully won't break too much).

In other news I added methods for getting all kinds of stuff from
warping and boundary conditions from the HDF5.

I also changed where some global constants in the WepyHDF5 module for
defining the records are used. Before they were coded into the
function and now they are constants that are used.

Going through this process made me realize that there was an
incongruency between what was determined by WepyHDF5 when making
records and what was expected from the boundary condition and warping
records in the boundary conditions class.

I was passing the walker index back with warp records but not the
cycle idx. But it was hardcoded into WepyHDF5 that the walker idx was
needed.

So I guess I am doing the same thing in effect but I made all that
happen WepyHDF5 and instead of using the walker_idx in the record I am
giving back I just write a 'record' for every walker, and put a None
where there was no warping. That sort of mimics the resampling records
so it makes sense, I guess. And sort of decouples them, which is the
goal!

I also implemented the actual boundary condition records and made a
trivial one for the unbinding one where it just makes a record on the
first cycle which is just the cutoff distance.

I also removed aux data for the cycle and time, because the cycle
should be in record and the time was useless because it was just some
time interior to the openmm context so it was always the same.

One last thing I know needs to be done with the boundary conditions is
that it should have mdtraj removed! but that is for another day.

Cheers