returning the resampler records from resample method, WExplore1
This involved restructuring everything in a slightly nicer (but still
mixed model with global state and functional) way.

So now there acutally is an assign (or score method if you like) and a
decide method.

This is accepted from the sim manager but not passed to the reporters
as of now