mocks for init system, run fields system growing
I unified the fields as: names, shapes, and dtypes in unbinding and
wexplore1

there are methods to get them in the superclasses as well, and
overrides in wexplore.

Methods for init in HDF5 just use the types for the groups to
initialize them and call these new fields methods.

Removed all aux data types since all data for records etc. is now
treated all the same.

Added a 'progress' group to the boundary condtions for the continual
data like min distances.