resampling_records in new format
Working on mocks to get resampling records data in the new format from
the WExplore1 resampler.

This involved changing how the decide and assign functions returned
data.

Redid the fields etx for the clone merge decision class.

Abstract definition of decision_id and target idxs. Users of this can
add this to their fields. In this case I add the assignments. Removed
the unnecessary methods there too.

Updated records to make it simpler.