mocks for reporting, fixes to dtypes etc.
Lots of little typos and discrepencies that I got rid of and should
really be made into constants but I was too lazy right now.

Everything is now returned as feature vectors to the reporter.

Resampling is now sporadic because I want to keep track of the cycle
because I changed it to having one record per walker per cycle. So
that there are multiple records per cycle. The step index and walker
index are also put into the record dict in the resampler.

Methods for reporting them makes it cleaner and easier to adapt in the
future.

Image dtype and shapes are now None because they are unspecified in
the source code and must either be given by the distance metric
somehow (don't want to do this, because it is a pain in the ass to
implement that everytime in the distance metric, but it is a
possibility to have everything perfectly lined up at the beginning) or
the way which I want to do it which is to take the first image
returned and initialize the dataset as that.