WExplore gets image shapes and dtypes automatically
THis is an API break for WExplore1, you now need to give it an init
state on construction. This is used to initialize the region tree and
to get an example of an image so it can provide teh shape and dtype
for HDF5.

Also turned all the field methods to normal methods.

renamed append group records to extend for HDF5