resampling, resampler, and warping records done
These are now working.

To do this I had to add back in the saving of the decision enum into
the file. This goes into the run as 'decision' and basically is the
same as before.

To be able to get reasonable records that don't have large data (such
as the images), I implemented a subsystem for specifying which of the
data fields are teh record fields. This has to be implemented for each
class that is going to be reporting data.

These fields are saved as a string dataset in _settings.

Then when records is called only those are used to make records.