analysis examples for lennard jones
This required some fixes to some getter type methods

Namely I had to change the way in which field data was prepared for
going into the records. This involved checking first whether the data
is variable length then if it is too many dimensions, and then whether
or not it is just one value or not.

This was needed because when you had a variable length dataset it
would just see how big that individual vector was then either tuple it
or take the element. When there was only one it would just take it,
then this would give nontuples for e.g. target_idxs for most of them
and screw up the rest of the code using those.

then ther ewas a small error in the way that the transition matrices
were made