REVO now saves images, must give init_state in construction
This is similar to how WExplore got this capability.

Except because a record in REVO is only once per cycle we have images
for every walker in a single record and so it had to be linearized
like the distance matrix so we also store the shape of the image as a
resampler data so it can be reshaped afterwards.