bug fix UnbindingBC, testing of restart reporter
There was a bug effecting the restart reporter in the UnbindingBC
method for getting the bc record fields where a zip object was not
being listified. This was fixed and the restart reporter just worked.

I tested the restart reporter with the Lennard Jones stuff too.

Did some other minor adjustments