restart sim manager, added continuation data to HDF5
Added a method to restart a simulation from the simulation
manager. This really just passes a run index to the reporters on the
init() call.

Then the HDF5 reporter knows how to handle this.

I also updated teh CLI for the we_Restart.py scripts so that you can
say which one you want to restart etc.