reworked code to be clearer and less nested
I moved the check to see if this cycle is the last one to outside of
the main loop and just have it run a continue if this is found to be a
warping record on the last cycle.

The comment for that line was also misleading so I rewrote it.

I also reworded some of the other comments to make it clearer why the
constants are used.

I also cleaned up the whitespace.