cleaned out constructor, added clone file
I got rid of a lot of unnecessary variables, some that literallly
weren't being used anymore and some I factored out by using methods on
the HDF5 dataset, so now there is no state the object has that is not
from the file.

I did this in working towards being able to clone a WepyHDF5 file
without the run data.

Which I wrote using the h5o.copy method