different file restarts
working on being able to make a continuation from one run into another
file.

This is useful for a number of things.

In any case there were lots of things to fix and improve with
restarts.

In dashboard I added an init method to make sure it is in the right
mode to overwrite dasshboards after being through a restart

Made some attributes in the hdf5 reporter hidden to cleanup the
interface, this wasn't complete.

Big changes to the `new_sim_manager` method where you can customize
the path automatically for reporters so that they don't overwrite the
first run's files.

These options are to either/and add a suffix to the filename and to
change the path to that file. So you can't change the name
completely. This is because you would have to name each reporters file
the same, so there is no way to do that automatically.