modified record access methods for contigs
We made new methods for getting contig records (as opposed to run
records) but really just reworked all the underlying functions to
work over contigs as opposed to single runs.

THen the run records call just calls those with a single run index.

I think I fixed a bug in the continual records method.

The named record methods were all converted to default to accept a
list of run idxs by default, so as to not duplicate methods too much.

All this needs to be tested