fixed link_file_runs, checks to continue_run
Fixed an error in the link_file_runs function where the wepy_h5 needed
to be open. It opens it in a context manager now before performing its
task.

Also previously creating a new run in WepyHDF5 with a continue run
didn't check to see if that was even a valid run in the file. Now it
does and raises a ValueError