cycle_tree, cycle_tree resampling panel
Wokring on the cycle tree needed to get branching parent tables so we
can get transitions from entire contig trees.

This led to me redefining the parent_panel methods in the decision
interface.

Really we only need the method that takes a single step of resampling
and returns the parent indices of the children that would result from
it.

this new method then is `parents`