cleanup, contig sliding windows
Got to the method for actually producing sliding windows over contigs.

This led to going back and cleaning up some of the other methods and
rearranging them.

Everything is kind of centered around the method that basically is
ancestors for the contigs.

Then from that you can generate a single parent table.

I also finished the algorithm for getting the tree leaves using a
recursive strategy.

Once you have the leaves then you can use the individual contig method
to get the spanning contigs.

THose are used to then derive the spanning contigs.

Still a WIP and working on a test case.