working on fixing and redesigning clone/merge subroutines WExplore
Working to diagnose the problems so there is some debugging stuff in
here.

Otherwise working on cleaning up and isolating object side effects in
the region tree and WExploreResampler object and to clean up and fix
problems in the subroutines for deciding the clones and merges for
each leaf.

Made a system to allow for different methods of figuring merges out,
since it is probably an NP-Complete problem.