next first draft of working WExplore
From my tests in randomwalk there are no errors and it seems to
replicate the original results.

In this version of WExplore there is a new debug mode that allows for
some special facilities in debugging what goes wrong in a WExplore
simulation.

The main system that allowed for the identification of errors was that
which checks the clone-merge specifications (merge groups and number
of clones per walker) and tells you which walker specifically was the
culprit in producing illegal children.

This of course would not be possible with the previous system for pmin
in which it was not actually enforced strictly.

There is a mechanism for "replaying" a resampling step when such an
error is detected in order that you can specifically use the debugger
to go back using a cached copy of the unmodified region tree and
initial walkers. This will then exit the program.

This wqas used with great success to fix two remaining very tricky
bugs that were detected from the randomwalk example with the number of
dimensions set at 5.

I am not guaranteeing there won't be other errors but now with this
mechanism hopefully a large class of them will be caught.