working on orchestrator
had to strip down a lot of stuff. No more apparatuses.

This is all due to the fact that roundtripping of hashes isn't stable
due to the fact that some classes when read back in from dill with the
class included cannot be encoded to the same thing.

I don't know why...

The CLI is working now for reconcile and running things as well.

The major implication is that the hashes have to be taken for granted.

That is they are now nominal and may not actually match the hash of
the objects encoding anymore :(

We can still do everything the way we want to except the identity of
an object cannot be known from this now.