updated reporters to kwargs model
In this commit I transition over to a more managed way to send
"reports" from the sim_manager to the reporters.

The sim_manager collects variables into a dictionary and passes them
as kwargs to the reporters. We keep a list of the keys that the
sim_manager will pass as a constant so that you can have dynamic
accepting of parameters by looking at the properties of the
sim_manager.

The sim_manager will check itself against this constant to make sure
it has all of the correct keys in its payload report before sending it
to reporters.

The reporters themselves also have a list of keys that they need to
accept.

This makes the task of using the keys a little tedious in the report
method, but it allows for some static checking to make sure the report
has the necessary data, which is good for debugging.

Also it clarifies the interface if a reporter is built that needs new
data the sim_manager will need to be updated to give that information
as well.

I converted the two main reporters to this format.

I also added skeletons for the new reporters I will do.