reporter filepath names do intelligent renaming if duplicates
If there are duplicate extensions in you reporters they will collide
when doing parametrization in the orchestration configuration.

So we add the class name into this to hopefully get rid of this.

The only case this won't work is that if you have two classes with the
same __name__ and the same extensions. Which I'm not sure can even
happen.

THis also means that every orchestration ready reporter has to make a
tuple of extensions instead of a single one.