proper multifile reporter support in orchestrator, image reporter
Previously I supported multiple filepaths in a FileReporter but I
didn't really do it properly.

I now have good handling of this.

First you can define explicit keywords for the file paths you support
in a class constant. THe order of which matters and is the order they
are stored underneath in the 'file_paths' attribute.

So you can either pass them in as explicit kwargs or just pass them in
as file_paths (although they will need to be in the correct order
which is annoying which is why the explicit keywords are used).

However, for automatically producing file paths (for instance in the
configuration class in orchestration) that is useful. And that now
supports that.

I also added a mechanism to make sure the filenames don't collide
between reporters.

Additionally you can specify (and must right now) the suggested
extensions for them for the configuration to do filename
parametrization correctly.

Made the WalkerReporter work using this mechanism. It was the
motivation since we need the topology file and the trajectory for
visualizing etc.

Also added the reporter for WExplore images that are just subsets of atoms.