LayoutGraph, Contig class, WepyHDF5.get_contig_trace_fields + clean
Kind of big commit.

I separted the functions that the ParentForest was starting to take
in.

That is I made another LayoutGraph class that has the base methods for
making layouts on graphs.

A lot of the methods for setting gexf attribtues etc. are messy and
not specially relevant to a ParentForest which can be used for pure
analysis, and will be in the future. SO it was prudent to clean it out
and make visualization separate.

This LayoutGraph class holds two networkx graphs. The one you feed
it (which in my examples is a ParentForest) and a special viz graph
which is used to hold the attributes for visualization. THis keeps the
namespaces separate for attributes and keeps a clean one for data.

You do duplicate the orginal graph but IDGAF.

I changed over what a ParentForest has to start it from a parent table
to a Contig, which is a new class next to ContigTree that represents
just one contig.

The goal of this was to offload some of the methods in the WepyHDF5 to
another level of abstraction (the contig that is) while being able to
access things intuitively from the HDF5 itself. That is fields and
records from across multiple runs. The WepyHDF5 had this but they were
based on lists of tuples (blhablahblha) and was confusing. I think
specializing the WepyHDF5 class more is good because its already
overloaded!!

Now to implement this I just subclass the ContigTree and do a check to
make sure the shit its given does not form a tree (for which you would
use the contig tree). But everything else is the same, except that it
adds new attributes like the contig_trace, and new methods for getting
fields etc. that don't require getting a list of tuples beforehand.

This makes it super easy and intuitive to use the contig. Basically
you use it just like you would a run in the WepyHDF5 and it takes care
of everything.

Now implementation wise some of the contigy methods are still in the
WepyHDF5 (because I didn't want to lose focus and time on those) but
the contig interface is there and ready to take over when the time is
reight.

Furthermore, the contig defines an object that is better for creating
the ParentForest so that there is now ambiguity about the correctness
of your parent table. Which is continually an issue for people who
like to mangle and dredge up zombies to pass into my stuff, and then
complain it doesn't work.

Now if you use the class it will work and someone complains I'll know
its actually my fault.

As far as stuff in WepyHDF5 I did a bunch of cleaning of of old contig
stuff that wasn't needed in it.

Furthermore I removed the spanning_contigs from it and put them into
the ContigTree where they belong and they produce real contig traces
and not lists of runs.

Lastly I implemented a missing functionality in WepyHDF5 which is to
get field data based on a contig trace. While this differs from the
interface for the records (which goes by just run_idxs) it does match
it in scope, which is why I implemented it in WepyHDF5 and not in
Contig itself which is where the me talking earlier in this message
would have said it belongs.

The reason for this inconsistency is that I have an immediate goal I
am trying to get (making tree layouts for a reporter so that I can
start simulations) and if I lose track I will go down the rabbit
hole. And furthermore it is kind of an implementation detail. And I am
not worried about deprecating all the contig methods in WepyHDF5 in
the future.

No one is going to be able to figure this shit out without
documentation anyhow.