new tree algorithm that eliminates tangles
This algorithm works by moving all the nodes to either side of a new
node that is being placed in the next step.

That way you never have crossing over!

It looks purdy.

I think there might be an unnecessary level of the big nodes still and
I might want to get rid of it. but in any case it works