finished ResTreeReporter, centering and row spacing in tree layout
Also moved the resampling_panel staticmethod from WepyHDF5 to the
parents module.

Maybe it doesn't belong there. BUt I definitely needed it for
iteratively making a parent table in the ResTreeReporter

That made me realize that most of the stuff related to making the
records was only in the WepyHDF5 class.

Perhaps it should find itself a home so it can be used by all. I am
just not sure where.

It wasn't too hard to re-implement the important bits for the
ResTreeReporter because most of the complexity in the WepyHDF5 is
getting it from its own stores and dealing with sparsity etc.