made a new subset function that preserves ordering
This will go to make subsetting topologies much less hard.

Before you had to get the order right and keep that the same
everywhere.

Now you just need to use the same order. Before when you subset it
would scramble the order of the atoms.

So now it is much easier to get subsets of atoms as PDBs.

Furthermore I was lucky enought o have written the things in geomm
well enough that they don't care about order. Which is a good thing.