switched order of arguments on WepyHDF5 constructor.
This is to support the use of the second argument that looks like
positional like the python open method

In python you can say: open('file', 'r') as shorthand for
open('file', mode='r') which is very common.

Before this wouldn't work since the second keyword argument was
topology. With this change you can write WepyHDF5('file', 'r').

And as always explicit is better than implicit but this is a good safe
guard and uses the principal of least surprise