Remove per frame arguments for traj_fields_map related methods
body='''
In the last implementation we allowed for passing either single values
to be shuffled on to the function to be mapped from the
traj_fields_map function, this would then expand them so that there
was a tuple to pass to each invocation of them. THis was following a
mapping API. Other than this choice you had the option to pass
separate arguments for each invocation in instaed of it automatically
being done for you.

I don't think anyone ever used this function and I just realized that
it prohibits the use of single arguments that themselves are
sequences (i.e. list of atom indices).

So I am removing the ability to pass individual arguments
in. Extra arguments will be used for each call.

This shouldn't effect anyone since I doubt anyone was using it like
this anyhow.

If you want that pattern (and really why would you since time
dependent stuff should be done with contiguous trajectories frmo
traces) you will have to do it yourself. '''

changes = [:changes/stricter, :changes/fixation]
domain = [:domain/src, :domain/docs]