Add json_top module with introspection functions
body="""
There was a need for more facilities to interact with and manipulate
the JSON topology objects themselves directly.

An increased distrust of external packages like MDtraj has increased
significant mental overhead when doing simple decisions due to the way
in which transformations to the data are done during various parsing
routines.

Additionally, the desire to select atom indices based on atom or
residue properties without using mdtraj was desired. This is because
the DSl-eque emthods in MDtraj to get information is highly unreliable
and tied to an internal dictionary and informal naming scheme to
accomplish tasks like `is_protein`. I would rather see and manipulate
the actual information for eafch atom and residue and make those
decisions, even though they are less declarative and simple to
express.

Normally to do this in MDTraj I would produce a pandas dataframe to do
this and look at the data since that is a reasonable tool that is
agnostic to the actual contents of the data.

Instead of going through mdtraj to do this, I elected to make those
functions available without it. These perform no reindexing etc. so
they treat the data as not volatile. Thus requiring little overhead.

I also made different representations for chain, residue, and atom
based views with keys to the other collections.

This is a small amount of code and meant to be specific to the JSON
representation. Better representations are still desired but for now
this is what we have.

Also moved the json top subset to the new wepy.util.json_top module
since it more rightly belongs there. SO taht is a breaking API
change. The necesssary internal callas to it have been refactored.

"""

changes = [:changes/accretion, :changes/rename, :changes/accretion]
domain = [:domain/src]