adding fields updates fields in place
[body]

When you tried to overwrite a field it was just calling create_dataset
which would fail if the field already exists.

This now uses the require_dataset and creates it if need be.

THis is run in exact mode so that the data is the same when you add
and raises a special error when it is not.

changes = [:changes/fixation]