Adds default SWMR support for WepyHDF5 reporting
[body]

SWMR (single-write-multiple-reader) is a feature in HDF5 that allows
readers to get a consistent view of the data while a writer is
writing.

This has the added bonus of making writes more ACID.

Both of which (although the latter is more pressing) are both good
features for wepy.

Thus HDF5 1.10 must be used (although I don't anticipate that it will
be much of an issue given the user base).

The only repercussion is that now you cannot dynamically add
trajectory fields. They must be specified up front and the datasets
for them must be created at WepyHDF5Reporter.init().

A warning is raised if a dynamic trajectory field is given and not
specified.

So in that sense it is a breakage (but really it is best practice to
not do this and I shouldn't have gone so out of my way to add this in
the first place, which just allows bad practices to persist).

But it is a growth item because it is increasing the reliability of
the code base in the face of faulty hardware. I am calling this a
'reliability' item since it is not a "bug" or error in the code but
rather a feature that makes the code more robust to the underlying
system.

To avoid the breakage and regression in performance the use of SWMR
mode is optional, so we are just left with a regression in
noisiness (the extra warning) and cognitive overhead in having to
manually configure something.

changes = [:growth/accretion, :growth/robustness, :regression/noisier]
domain = [:domain/src]