adds cycle idxs to orchestrator run data
[body]

This was necessary to get proper reconciliation of runs from
checkpoints.

This is because the actual HDF5 file may contain more data than what
was saved from a checkpoint.

So we needed to add a mechanism for saving the last cycle idx of a run
in the orchestrator for the runs.

The only external API that is changing is the orch.register_run
function which now requires a cycle index, so that can be considered a
breakage.

changes = [:changes/fixation, :changes/accretion, :changes/stricter]
domains = [:domain/src. :domain/docs]