removes namedtuple classes from restree reporter attrs
[body]

This was originally used to be able to generate records on the
fly. However, there are issues with pickling them now so they need to
just be generated temporarily. It is a good practice.

Requires patching of old orchestrators to rid them of those
attributes.

changes = [:changes/fixation]
domains = [:domain/src]