fixes bug in combine_orch_wepy_hdf5s
[body]

When you had runs that continued each other you would get duplicate
calls to set the hash for a run (at the end of one run and the
beginning of another). So I just made an exception for the
AttributeError and continued if it fails.

changes = [:changes/fixation]
domains = [:domains/src]