adds support for num_walkers
[body]

I needed to be able to get the number of walkers for calculating total
sampling time in the rates calculations so I added the methods and
interfaces for doing this in general in WepyHDF5 and the contig. Even
though in WepyHDF5 we don't support different number of walkers yet, I
made it so that it provides that interface that other applications can
build against and will have once that feature is made available.

This does rename the n_cycles to num_cycles in contig because I am
trying to move everything in wepy to a num for getting numbers like I
did in the WepyHDF5.

changes = [:changes/accretion, :changes/rename]
domains = [:domain/src]