adds WepyHDF5.copy_run_slice method
[body]

I needed to be able to slice runs from one file and copy them to
another and pretend that it is a full run. Mainly this is for
truncation with orchestrator checkpoints, where the HDF5 run has more
data than the chckpoint has and so getting contigs from orchestrator
checkpoint runs would have been wrong.

I wanted to do something fully-featured with RunHDF5 classes etc. but
it was getting complex really quick. So I opted for this kind of
minimal solution that does some of the hardest bits of taking "slices"
across a whole run including trajectories and records.

changes = [:changes/accretion]
domains = [:domains/src]