adds the kv-store to utilities
[body]

I want to start using SQLite3 for the container format for
orchestrators but I wanted something simpler than the full relational
database since I will just be storing simple kinds of values.

I also wanted it to be portable so I figured this is the easiest way.

I just copied the code from a well-known key-value store for SQLite3
in pure python to our project so we don't have another dependency, and
since it is so simple.

changes = [:changes/accretion]
domain = [:domains/src]