modifies kv store to just store bytes
[body]

The module I copied did things via JSON, which really isn't what I
need since I will be storing a lot of pickles, so I just simplified it
and made it do no preprocessing with JSON and just stores whatever you
give it as long as it is bytes or bytearray.

Nothing could possibly depend on this so I am eliding the breakage keywords.

changes = [:changes/fixation]