reimplimenting orchestrator and snapshots
[body]

This is the first major round of changes and is not complete.

The parts that are working are the Orchestrator (currently the new
implementation is under NewOrchestrator), and the methods for storing
snapshots, configurations, and metadata like defaults.

That leaves the run information table which is going not going to be a
KV-store.

WIP = true