updates run methods to new implementation
[body]

Reconfigured things so that the input orchestrator will only ever be
read from so there is no need to copy and modify it in place anymore.

Furthermore, the protocol for saving checkpoints should be faster
since we now just have to serialize the intermediate snapshot.

The update of the single checkpoint orch database doesn't require
serializing the whole thing and so we can update it in place. FOr this
I modified the _save_checkpoint method to do this in a way that is
completely transactional (I assume using the write-ahead log of
SQLite) and adds the new snapshot, removes the old one, removes the
old run record and adds the new one.

This is important since we are not just serializing the whole
strucutre and so we have to worry about data integrity. This solves
that problem.

The _gen_checkpoint is gone now since it really wasn't necessary.

Also there is another command that must be run to initialize the
database at the beginning of the run for the checkpoint
_init_checkpoint_db.

The other significant thing to note which won't change the CLI
interface inputs, but does change the way that files are outputted is
that the final orch file won't be craeted. This is a good thing. In
the past it was a highly inneficient thing that basically duplicated
the entire input orchestrator causing huge bloat. Instead the final
snapshot is written to the checkpoint.

So the checkpoint file (checkpoint.chk, which should be changed to
something better) is the only output containing snapshots. It contains
2 snapshots (start and end) at any time, a single configuration, no
metadata, and the run table (with 1 run in it). SO it is decently
lightweight.

Any other approach would require access to the main orchestrator
database, which causes a huge operational problem because potentially
many simulations will be needing access. This would require a
significantly more complex system for managing concurrent connections
that would probably end up causing a lot of problems (mainly because
if a process blocks and then fails (very likely) it would halt all
progress for the other simulations).

So I think it is a good tradeoff.

Still need to clear up some stuff, test it, and clean up the
documentation so still WIP.

WIP = true