updates the layout of the project, packaging, testing
[body]

I wanted to add tests and so I wanted to switch to the:

- src
  - wepy
    - module files...

Which is the suggested way of doing things.

So I did that and wrote a minimal sanity check for the testing suite
to run a test on nothing.

Also got tox up and running for multiple python versions. This is easy
from the configuration point of view in the project but requires the
installation of multiple different versions of python. I solved this
in a script taht automatically installs them to a place so that tox
can use them. This is outlined in the readme and how to run it with
those pythons.

Also added various other project files like authors and MANIFEST.

Started using the requirements.txt file because I know understand the
difference between abstract and concrete dependency specs (discussed
in readme).

This allowed me to fork and patch the setup.py in mdtraj for
installation from my git repository so that it happens automatically.

Added mdtraj as an extra install that installs it with a separate
option. See readme