fixes DB URI generation; adds append_only mode
[body]

- allow for better URI+mode handling, allows for shared in-memory
  sessions by default

- implement append-only mode for KVs and Orch

This is an important one as I was trying to figure out how we could
reliably be able to read from a master orchestrator at runtime for
simulations to get the snapshots they need. Because a DB is locked for
writes, if I was reconciling things to the master orchestrator while a
simulatino needed a snapshot it would fail and drop the job. The
solution is to reduce the isolation level by making a connection in
read_uncommited mode which allows for reading even while there are
write locks. This means you can potentially have dirty reads, but
since really our data stores are append only most of the time this is
okay. To make this concrete and safer I added specifically the option
for this that must be turned on.

- update tests submodule

WIP = true