adds new subcommands and redoes command line
[body]

With the orchestration overhaul it is now practical to run simulations
from standalone snapshots in pickles. So I made a mode for doing that.

I also added more subgroups for commands for ls, and get. THe get is
new and facilitates the running from snapshots usage. THis gets a
snapshot from an orchestrator and writes it to a pickle on the
filesystem. There is now the corresponding get and ls commands for
runs and configurations as well, which is made easily by the
subgroups.

changes = [:changes/accretion, :changes/breakage, :changes/rename, :changes/removal]